+++ packages/moonstone/CHANGELOG.md
@@ -30,6 +30,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/ExpandableItem` to correct the `titleIcon` when using `open` and `disabled`
 - `moonstone/GridListImageItem` to center its selection icon on the image instead of the item
 - `moonstone/Input` to have correct `Tooltip` position in `RTL`
+- `moonstone/VideoPlayer` to fast forward properly when video is at paused state
 - `moonstone/VideoPlayer` to correctly change sources
 - `moonstone/VideoPlayer` to show or hide feedback tooltip properly